+++ CHANGELOG.md
@@ -8,8 +8,13 @@
 ### Added
 - setup script
 - added customization for on_message
+- error messages when user is not a mod and the user runs commands that requires being a mod
+### Changed
+- `/reload` now requires full path and can access `ext`
+- when add moderators embed will now be outputed instead
 ### Fixed
 - can force terminate now
+- settings didn't work when there are no prefix except mentioning
 
 ## [2.2.1] - 2020-09-01
 ### Optimized